# jdbc backend: stream osh cells for the given cell id range