The old implementation always loads all osh cells for the given cell id range into in a static collection before streaming them out (to the parallel stream that invokes the CellIterator, etc.).

The new implementation makes loads the data as needed by converting the jdbc `ResultSet` into an `Iterable` and streams that (as outlined here: https://dzone.com/articles/adding-java-8-lambda-goodness). :sunglasses: