@tyrasd 2018-08-30
Forgot to post the benchmark results:

```
# before

building histocounts: 2.93s ±0.25s
highway histocounts (length): 3.21s ±0.17s
forest histocounts (area): 7.67s ±0.21s

# after

building histocounts: 2.70s ±0.16s
highway histocounts (length): 2.68s ±0.16s
forest histocounts (area): 7.14s ±0.22s
```

Overall this even results in a slight performance improvement.